refactor: 구간들 (Sections), 테스트 코드 중복 제거 리팩토링
